test CarNameCheck, TryCountCheck 객체가 데이터 체크 책임으로 변함으로 인하여 테스트 코드 수정
